add suggested renaming for test utils
